Update patch set 2
Patch Set 2:

Hi Andrew,

Re: check for IQMakeProjectInfoListener:
Actually there is no check on QMakeProjectInfo side. The reason is that I wanted to notify listeners without 'synchronized' section to prevent possible deadlocks. The check can be placed there only if we would notify listeners within 'synchronized' section. One of the solutions might be to describe this behavior in JavaDoc.

Re: not passing data to listener:
Reason for this way that we wanted to have it done lazily. If we would be passing data to listener event then we would need to calculate a new QMakeInfo information and that would need to be calculated in a thread request a change of QMakeInfo on QMakeProjectInfo e.g. ResourceChangeListener.resourceChanged(). We do not want to block this thread. In fact, e.g. QML indexer is just reacting on the qmakeChanged() via scheduling a background task that asks for recalculation of QMakeInfo data... So for QML-editor, this behavior is works well.

Re: change Notify-1 vs. Notify-2:
Actually we expect that the user of IQMakeProjectInfoListener is having old QMakeInfo value stored in its structure and a new QMakeInfo value is retrieved from IQMakeProjectInfo.getActualInfo() - QML editor does not care about mid-states since it would need to recalculate it again anyway.

Still we are using this feature for QML editor only. Certainly there might be other use-cases too. If you have any, please, let me know and we will extend the API to cover those as well.

Patch-set: 2
Reviewer: Gerrit User 1821 <1821@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0