Update patch set 2
Patch Set 2:

I still think this change is OK to commit now.

However, I'm not sure that completely described my thought regarding the check.  It is sort of irrelevant now that I've looked at the existing implementation to see that it behaves properly.  However, it might still be worth adding a comment so that future implementers of IQMakeProjectInfoListener#qmakeInfoChanged do not create hard to debug bugs.

This is the specific scenario that I think needs to be avoided:

public class L implements IQMakeProjectInfoListener {
    private IQMakeProjectInfo qmakeProjectInfo;
    private Object obj;

    public void register() {
        obj = new Object();
        qmakeProjectInfo.addListener(this);
    }

    public void destroy() {
        qmakeProjectInfo.removeListener(this);
        obj = null;
    }

    @Override
    public void qmakeInfoChanged() {
        // *** This will be null in the rare case where the listener is being
        //     notified near the same time that #destroy is called.
        obj.toString();
    }
}

The NPE would depend on timing and would therefore be difficult to debug.  Implementers of IQMakeProjectInfoListener#qmakeInfoChanged should be aware that calling remove does not mean that they will never be notified.

Patch-set: 2