Update patch set 3
Patch Set 3:

Hi Andrew,

Agree that this is a race-condition that is implied by the implementation of QMakeProjectInfo. I agree that we need to extend JavaDoc for IQMakeProjectInfoListener.qmakeChanged() method at least.

The only other solution is to fire notification under 'synchronized' block. I am not sure if this would be good since we would expose that 'sync' from the implementation and potentially lead to possible deadlocks.

Patch-set: 3