Update patch set 6
Patch Set 6:

> If this is the case it would not work as the current structure is set to use one Retrieval per Memory context, the UI based the structure on this e.g. the Memory Browser uses one Retrieval per Tab. Switching Retrievals per memory context fixes the multi-process memory context for both the Memory browser and Memory View which rely on this relationship.

Answer: Not sure I understand it. Currently there is only one instance of the memory retrieval per session, right? 
I am suggesting to replace it by a new implementation of the "IMemoryBlockRetrivalExtension" interface that would implement internally the same functionality as the retrieval manager. It will maintain the list of context specific retrievals and the "context" -> "memory retrieval" mapping. When the "getExtendedMemoryBlock()" is called for some context this class would check if the corresponding retrieval is already created and call its "getExtendedMemoryBlock()". If there is no retrieval for the given context it will create a new one. Does it make sense?

Patch-set: 6