Update patch set 6
Patch Set 6:

Yes, the idea of using a single IMemoryBlockRetrievalExtended which would delegate to the real DsfMemoryRetrieval makes sense.

The problem is that we actually want the views to know there are different MemoryRetrievals.  We didn't explain this anywhere outside of the code, so I can understand it is not obvious.

Currently the memory monitors in the Memory View are only applicable to the debug session for which they were created.  That means that if you run two debug sessions, monitors of one session are only shown for that session.  When you select the other session, the previous monitors disappear and you can create new ones for the second session.  When switching between sessions, the relevant monitors appear while the old ones disappear.  This makes sense because a memory address for one session may not be applicable for a different session.

Both the MemoryView and the MemoryBrowser behave like this and achieve this by associating memory monitors (or memory tabs for the MemBrowser) to a particuler MemRetrieval.  When selecting a new session, a new MemRetrieval is used, and the views know to change their list of monitors.

We want the same behaviour when switching between two processes (or I should say IMemoryDMC).  The monitors created for one process (IMemoryDMC) are not necessarily applicable to another process of the same session.

To be able to get the views to do that, we need to use different MemRetrieval explicitly.  The Mem views notice a new retrieval for a different IMemDMC and use new monitors/tabs.

One of the original ideas for this 'feature' was to use a single retrieval and let the views figure out when the IMemDMC was different.  This turned out to be very complicated for the MemBrowser and not really feasible for the Memory view without changing things in the platform.

On the other hand, exposing different retrievals allows the views to work as we want without any changes.

Sorry for not explaining this before.

Patch-set: 6