Bug 424499: Find References does not work for Qt signals and slots
The QtASTVisitor is trying to use the QtIndex during indexing.  Any
results available at this time are based on the state of the AST the
last time the code was indexed.

This adds a test case to reproduce the problem.  The test cases indexes
the project one time.  It should find two references to the signal.  If
the QtIndex data is stale, then it will find 0 references.

This also replaces the code that looks for QObject::connect function
calls.  The proper behaviour is to find all overloads of #connect as
well as references with QObject::disconnect (all overloads) function
calls.

A new test case checks for references in all overloads of #connect and
#disconnect function calls.

Change-Id: I28fc4213d6dddff10f81a6bd3ef01e24c74f31db
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>