Bug 422841: Add Q_GADGET to QtIndex
This extends the QtIndex to support C++ classes that have been annotated
with the Q_GADGET macro.  QGadgets are normal C++ classes that are able
to host Q_ENUMs.

The implementation classes for QObject have been modified to share
common parts with the implementation for QGadget.  The types are
intentionally not related in the QtIndex API.  This allows for divergent
changes in the Qt spec.

This patch includes new tests cases for Q_GADGET.

Change-Id: I59eb745ff5614c2897d67dd7d6807763091120af
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>