Bug 425002 QObjects cannot be found within namespaces
This changes the Qt PDOM so that QObject's are stored using their fully
qualified name (including leading ::).  IQObject's API has been changed
to return the fully qualified name.  Leading :: is stripped from the
IQObject name so there is no change from the previous implementation for
names that are not inside a namspace.

This includes a new test case to check this fix.

Change-Id: I1786151463e9029cdf1f2c213466adc8c3aa3618
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>