Bug 422841 QtIndex API for qmlRegisterType function calls
The Qt spec includes a special function that is used to introduce C++
types to a namespace that is accessible from QML.  E.g.,

	qmlRegisterType<Q>( "uri", 1, 0, "QMLType" );

This will create a QML type called QMLType.  The type will include the
signals, slots, and invokable that are defined in the C++ class Q.  The
type is accessible in QML using the given URI and the version is 1.0.

More information is available at:

	http://qt-project.org/doc/qt-4.8/qdeclarativeengine.html

This patch adds IQmlRegisteredType, a collection of which can be
accessed from a new method in QtIndex.

This also includes new test cases for this feature.

Change-Id: I70c44d1d8d3a0594de44e692a16f7b26396e8464
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>
Reviewed-on: https://git.eclipse.org/r/20347
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>