Bug 425102 New check for Qt Codan checker
The Qt Codan checker for QObject::connect function calls was not
confirming that the SIGNAL and SLOT expansion parameter were Qt signals
and slots.  The checker would allow the function call as long as the
expansion parameter resolved to a C++ method.

This patch changes the behaviour to required signals inside SIGNAL and
slots inside SLOT.

Change-Id: Ieec2f3a7ef4968d45ac3f6323b20c2f195fe3400
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>