Bug 425223 - [UI] Code Formatter should apply the active profile set in plugin_customization.ini
Added an initAndCheckVersion() method in org.eclipse.cdt.internal.ui.preferences.formatter.FormatterProfileStore class and it’s called by the start() method of org.eclipse.cdt.ui.CUIPlugin. This method is very similar with the checkCurrentOptionsVersion() method in FormatterProfileStore of JDT or JSDT. But it add a condition to check if the selected profile is included in the default build-in profile list. If it's not included in build-in profile list (e.g., it's defined in plugin_customization.ini like 'Momentics' profile), then apply it.

Line 86:
 if ((selectedProfile instanceof CustomProfile) || !profiles.contains(selectedProfile)) { // CustomProfile or Profile in plugin_customization.ini
 manager.commitChanges(instanceScope); // updates CDTCore options
 }

Signed-off-by: Rolland Liu <liurolland@gmail.com>