Bug 422841 QtIndex accessor for QmlRegistrations of an IQObject
This adds an API method to IQObject that will return the list of QML
registrations for that type.

I've also renamed the previous "QmlRegistered" to "QmlRegistration"
because it makes more sense.

The main part of this patch is a change to the QtPDOMLinkage.  It now
maintains an index of the registrations that are keyed by type.  The
index is a BTree that uses the QObject name as the key and a list of QML
registration PDOMNames as the value.

The list is created in the linkage's onCreate callback.  Old names are
removed in the list's onDelete callback.

This includes a test case for this scenario and also updates the
previous tests to match the new naming.

Change-Id: I4b994a51958c848b4e2a3209165e6e6866e22270
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>
Reviewed-on: https://git.eclipse.org/r/20527
Tested-by: Hudson CI
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>