Bug 425787: Reindex projects when Qt nature is added
If a project description is changed to add the Qt nature then the PDOM
needs to be rebuilt.

Since index rebuilds are potentially expensive, this first checks to
make sure the PDOM does not already contain the QtLinkage.  If the
linkage already exists, then it will be updated by the normal triggers.

The reindex operation should only be needed to add the linkage the first
time the nature is added.

This does not trigger a reindex if the nature is removed.  Without the
nature, the extra linkage will be safely ignored.  The C++ linkage is
(proportionally) much larger than the Qt linkage, so it doesn't make
sense to spend significant time rebuilding the index just for the small
space savings.

Change-Id: I263b05e4de407775979843f5d6a9c8c172948d72
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>