Bug 426238: Update head of external references list when needed
When a PDOMName is deleted and that name is the head of an external
references list, the list's head must be changed.

The Qt plugin is the only user of the external reference list.  One case
is the link from a SIGNAL or SLOT expansion to the C++ method binding
for the corresponding function.  In this case, the problem will appear
when all of the following are true:

1) The file containing the SIGNAL/SLOT expansion is changed and the
   index updated
2) The corresponding function is declared in a different file
3) The the function is the first entry in the external references list

When #2 is false, the binding (and the entire list) is removed as part
of updating the file containing the name.

When #3 is false, the list is updated with existing code using the
PDOMName's {next|prev}InBinding pointers.

Change-Id: I1e27c7c2356ca1fb68f57d69c40728289288ed66
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>