Bug 418406: Qt content assist is too aggressive
In Content Assist cases like:

    q->

(when the cursor is to the right of the arrow operator) the name that is
returned by the content assist context is the empty string.  This was
used to check if the QObject::connect function applied.  Since the empt
string matches all bindings the Qt assistant incorrectly continued.

I've added a check to stop processing when the current name is the empty
string.

Change-Id: I4e5bf52f4ca1ddc33d88a48917213adbbeb81836
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>