Terminal: Fix Bug 425757 - Command line is disarranged
This commit rolls back the commit
faa118886693ef2c44122fccbff137ba4670ed18 which changed the layout to use
a FillLayout. As this is fixing an issue with a horizontal scrollbar
painted on Linux where it is not needed, it leads to more serious issues
on Windows. Rolling back the commit is addressing the immediate need of
getting the layout fixed where the scrollbar issue is somewhat
cosmetical.