Bug 426648: Do not create 2.8 million base classes
This fixes a bug that was found while trying to get the indexer to
complete on the boost source code.  When doing a full index operation,
std/functional.hpp is included many times.  This file has a
specialization for a template class in line 31 (boost release 1.55.0):

    BOOST_TYPEOF_REGISTER_TEMPLATE(std::binary_negate, 1)

The template is specialized by each translation unit that includes the
file.  Each time that happened, the PDOMNode for that specialization
would double it's base class specifier list (it was adding all of itself
to itself).

The heap dump from my OOM had 2.8 million entries in the list (it ran
out of memory while trying to add even more).

This fixes the problem by not changing the existing specialization if it
already has any base classes.

I've also added a test case to catch the problem.

Change-Id: Ib0d487bc91ac1f0262d199b6851645458ad9b4cc
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>