Update patch set 14
Patch Set 14: Verified+1 IP-Clean+1

Patchset 14 are the last changes that I've made.  You can compare with patchset 13 to see the changes.

This is the functionality I want to commit, but I need to post one more change to deal with the fact that this patch will go into 8.4 instead of 8.3 which means new @since tags and plugin versions.

Patch-set: 14
Label: Verified=+1
Label: IP-Clean=+1