Update patch set 3
Patch Set 3:

Hi Olivier. I quickly tried the patch. One thing I think is missing is the ability to cancel the formatting and still have the document saved. Like in JDT, if you save and are not happy with the auto formatting, you can undo (Ctrl+Z) and the document will be saved in its original form. I'm not sure how this is done in JDT yet.

Patch-set: 3