Update patch set 7
Patch Set 7:

Thanks Marc-Andre for fixing the pom.xml version.

Patchset 7 add support for dprintf for the GDB Hardware launch.  For some reason the 'call' style which prints to stdout does not work in this mode.  I will investigate this further with GDB but for now, I set it to 'gdb' style, which will dprintf to the gdb console.

I'm having an issue with dprintf being set multiple times with multi-process.  This is bug 389945 which I will look at now.

Patch-set: 7