Update patch set 1
Patch Set 1:

My suggestion would be not to get rid of the code. I cannot point to a specific test case which would be failing - it would be nice to have one though. Nevertheless, the check for IType is too generic. When comparing the binding of the typedef/alias with the binding of the local definition of the class, which they are an alias for, the check returns true. Because at that point the visibility is still private, the completion does not work.

I would propose to check for ICompositeType instead of IType for both bindings and then cast to that type. Since we are in the context of a declaration specifier which is an ICPPASTCompositeTypeSpecifier, we can be pretty sure it will not break required functionality.

For verifying this I created this commit: https://git.eclipse.org/r/#/c/22233/

Can we agree on that?

Patch-set: 1
Reviewer: Gerrit User 661 <661@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: IP-Clean=0