Update patch set 2
Patch Set 2:

(2 comments)

I can shed some light on the code in doubt. I added this additional condition while developing the visibility for nested types: https://git.eclipse.org/r/#/c/11901/
In the second patch set the alternative to check for type equality was added. This was motivated by tests like IndexCPPBindingResolutionBugs.testAnonymousUnion_Bug216791. The anonymous union in this bug gets wrapped in a PDOMASTAdapter.AnonymousClassAdapter, for which the equals operation returns false. Subsequently, this test failed as the union in this wrapper had not been recognized as fitting member. During the further development of the feature I changed the aproach when the visibility is checked and therefore, which binding is used. 

In my opinion retrieving the visibility of the adapter does not need to work for the adapter, as it is queryable with the original binding. Or is this assumption wrong?

I'll submit an adapted commit.

Patch-set: 2
Label: Verified=0