Update patch set 1
Patch Set 1:

Hi Simon, thank you for the great work! Here are a few initial comments:

- Some copyrights are missing

- I think cdt.ui’s manifest.mf needs to be bumped to 5.8 and @since tags need to be 5.8.
 Make sure you set your baseline to the soon to be released CDT 8.3. You can get it at this update site: http://download.eclipse.org/tools/cdt/builds/8.3/nightly
The instructions about setting the API baseline are not updated yet since 8.3 is not released yet.

- The new refactorings seem to be missing from the main menu. But they are accessible from the context menu. I’m assuming the plugin.xml is missing the menu entries.

- I think Pull up and Pull down should be in the Outline view’s context menu when right-clicking on a member. Similar to Refactor > Rename and Source > Generate Getters and Setters.

- Inline should have a key binding similar to Java. I think on Windows/Linux it’s alt+shift+I. On Mac it should be option+command+I but it’s already bound to the include browser which I think is wrong but don’t worry about the Mac key binding, I can fix it in a separate commit.

I haven't really tested the refactorings yet or looked at the code thoroughly so expect more comments ;) I suggest you wait for more comments before pushing a new version of the patch to gerrit but that's up to you.

Patch-set: 1