Bug 425711: User preference to limit tokens per translation unit
This adds a new scalability preference that aborts parsing when too many
Tokens are created for a single translation unit.  This is a heuristic
that fairly close predicts the files that will be too complex for the
indexer to handle.

The token counter is disabled by default.

When it is enabled, the default token limit is 25,000,000.  This value
was determined by counting the number of Tokens produced for each
translation unit in the boost-1.55.0 sources:

sqlite> select * from counts where count > 10000000;
count       location
-----------------------------------------------------------------
100000001   libs/local_function/test/all_decl.cpp
100000001   libs/local_function/test/all_decl_seq.cpp
100000001   libs/local_function/test/all_decl_seq_nova.cpp
100000001   libs/preprocessor/doc/examples/array_arithmetic.c
99808587    libs/function_types/build/preprocess_arity_loops.cpp
62380381    libs/preprocessor/doc/examples/delay.c
58096841    libs/serialization/performance/xml/string256_test.cpp
58096828    libs/serialization/performance/xml/int256_test.cpp
52898416    libs/mpi/src/python/collectives.cpp
52573708    boost/spirit/home/support/char_encoding/ \
                  unicode/create_tables.cpp
21315014    libs/utility/binary_test.cpp
18799536    libs/math/test/test_rational_instances/ \
                 test_rational_double1.cpp
17758615    libs/mpl/test/string.cpp
13100401    libs/container/bench/bench_set.cpp
11976021    libs/local_function/example/const_block.cpp
11381198    libs/math/test/test_tr1.cpp
10432186    libs/parameter/test/preprocessor.cpp

This value means that the indexer will process all files in boost
without running out of memory on a 1Gb heap.

Change-Id: Ia9fc73dfb38454cc8735f537e3ac6e661864fb4f
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>
Reviewed-on: https://git.eclipse.org/r/22386
Reviewed-by: Doug Schaefer <dschaefer@qnx.com>
IP-Clean: Doug Schaefer <dschaefer@qnx.com>
Tested-by: Doug Schaefer <dschaefer@qnx.com>