Update patch set 3
Patch Set 3:

I was jumping back and forth to figure out when 'count' was in octets and when it was in addressable units.  So, I refactored all the names from count to word_count when appropriate as was done for some (but not all methods).  I also cleaned up MIMemory from all the useless comments.

I'll continue the review now.

Patch-set: 3