Update patch set 2
Patch Set 2:

(16 comments)

Thanks for the valuable feedback!

Regarding the distinction of C++11 attributes and GCC attributes: I left the GCC attributes as is and added the attribute specifiers for representing C++11 attributes. The distinction can be seen for example in the NodeWriter. If an I(CPP)ASTAttribute is encountered in a IASTAttributeOwner it is expected to be a GCC attribute. If an ICPPASTAttributeSpecifier is encountered there, it is expected to be a C++11 attribute. The previous implementation of attributes is a bit limiting regarding the capabilities required by C++11 attributes (scope and multiple attributes in one specifier).

I also thought about a solution which introduces an attribute specifier for GCC attributes, but that would break the current API. It's related to the problem mentioned in my comment (Line 23) in ICPPASTAttributeSpecifier.

Patch-set: 2
Label: Verified=0