Fix for deadlock when a WorkspaceJob notifies about Resource changes
A deadlock can happen if one thread is running WorkspaceJob that modifies
workspace resources. Then QMakeProjectInfo.updateActiveConfiguration()
is called - aka WorkspaceJob and QMakeProjectInfo locks are used.
If at the same time, second thread is using QMakeProjectInfo directly,
then it might ends up with calling
QMakeProjectInfo.updateActiveConfiguration() as well but this time
the locks might be used in inversed order aka QMakeProjectInfo and then
JobManager (from the call to CoreModel.getProjectDescription()).

The prevent the deadlock, QMakeProjectInfo needs to use the same order
of locks i.e.: WorkspaceJob and then QMakeProjectInfo.

Change-Id: Ic93e2353084b7a18767f339fa733d50ff8391e4e
Signed-off-by: David Kaspar <dkaspar@blackberry.com>