Bug 429196: stack overflow in IndexFileSet.containsNonLocalDeclaration
This fixes a StackOverflowError caused by too large of a search space in
IndexFileSet.containsNonLocalDeclaration.  The old implementation
searched through all fragments (.pdom files) in all of the index file
set.  It then eliminated bindings that were not contained in one of the
file set's PDOMFiles.

The problem is that searching the fragment leads to a searching the
entire index (in PDOMCPPLinkage.doAdaptBinding(PDOMNode, IBinding,
long[])) which leads back to searching the file set.

This changes reduces the scope of the search to only the PDOMFiles in
the index file set.  I think this is also more efficient because it is
not collecting many names only to eliminate the ones that are not part
of the fileset.

Change-Id: I8c4f7d859781e9b3c6e983efa6f570e1460e7a40
Signed-off-by: Andrew Eidsness <eclipse@jfront.com>