Update patch set 1
Patch Set 1:

(1 comment)

I added a comment about being more specific with the scheduling of the new runnable, but otherwise this seems to address the specific deadlock case that was observed.

What about the other places that sync is locked?  If I understand this class correctly, I think that calls through the process() method are OK, because it looks like it is only called when the proper locks are taken.  However, what about places like getActualInfo?

Overall I think the threading and lock requirements in this class are getting to be extremely subtle.  I wonder if there is some change that could be made to simplify this area?

Patch-set: 1