Update patch set 2
Patch Set 2:

As I pointed out in an earlier comment "I was wrong about the index dropping duplicate definition. I've made this mistake before and when I go back to check I remember that it is only in the AST (setDefinition overwrites the existing field), which is valid. The index does correctly handle duplicates in all cases I looked at."

Further, this patch is not fully correct, but it is wrong to say that it is worse that the current behaviour.

If there are two partial templ. specs right now, then the index will have the equivalent of:

    class Te<T*> : public B1, public B1 {};

Which is obviously non-sensical.  Why would a class have more than one of the same class as a base class.  This patch will create the more sane binding:

    class Te<T*> : public B1 {};

The right solution of having a binding with B2 as a base is missing from both solutions.

Patch-set: 2
Reviewer: Gerrit User 1042 <1042@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0