Update patch set 3
Patch Set 3:

I changed the current implementation to have a better distinction between C++ attribute specifiers and GNU attribute specifiers. I.e. I added an IASTAttributeSpecifier interface, which is derived by ICPPASTAttributeSpecifier and IGNUAttributeSpecifier. The IASTAttributeOwner now stores IASTAttributeSpecifiers.

It is not the final implementation as I need to have a look at the loction of the GNU attributes. They are not always the same as the C++ attributes. I think that can be handled in the AST writer though.

Patch-set: 3
Label: Verified=0