Update patch set 4
Patch Set 4:

It's ineviatble that the different attributes don't reside at the same location. E.g. attributed using directives:

* GNU: using namespace inner __attribute__((__strong__));
* C++11: [[attr]] using namespace NS;

Either we could separate the C++ and GNU attributes on the node (ASTAttributeOwner), or we introduce a filter to allow only certain attribute specifier types to be visited at the corresponding location.

The first option seems not that neat to me, as we would have a dependency on IASTAttributeOwner level to the more concrete concepts of GNU and C++ attributes.

I dont' expect there to be a similar problem with other nodes where we already solved this problem, right?

Patch-set: 4