Update patch set 4
Patch Set 4:

> We can extend IASTAttributeOwner to contain separate methods for accessing C++11 and GCC attributes and a convenience method for getting all attributes at once. Alternatively, the latter method can be external to IASTAttributeOwner.

Sure, I'll do that. Leaving the method for accessing all attributes is required anyway, as there are locations where both attribute types are allowed (I have to check the sequence GCC handles them), in which we need to retain the order as well.

> While working on attributes, could you please also take a look at https://bugs.eclipse.org/bugs/show_bug.cgi?id=413615. Thanks.

Of course, it looks like it will go along with improving the attribues anyway.

Patch-set: 4
Label: Verified=0