Update patch set 1
Patch Set 1:

Patch set 1 adds 2 new interfaces, IDisassembly3 & IInstructionWithRawOpcodes along with an extension of MIDisassembly which implements IDisassembly3. 
IDisassembly3  allows Opcodes to only be requested from GDB when required, any thoughts on if they should just be requested all the time for GDB 7.3<, thus potentially removing the need for IDisassembly3? At Renesas we have some components which request very large amounts of disassembly (possibly several MB) so wanted to avoid degrading performance when opcodes aren't required. Due to the implementation of the column opcodes are requested whether it is visible or not.

I have checked with GDB remote debug & local debug Windows (MinGW).

If this matters, Renesas does not currently have a company contribution agreement in place but is a Eclipse member.

Patch-set: 1
Label: Verified=0