Update patch set 2
Patch Set 2:

(1 comment)

I've added a comment around the #isTreeLocked line.

As for a different approach.  It seems that alot of the subtle behaviour is introduced by the decision to not pass a snapshot of the state as a parameter in the notification method.  I think the reasoning behind that was that you felt it was too expensive to construct the snapshot.  If you were able to come up with an efficient way to create the snapshot then you might be able to avoid these locking problems.

Patch-set: 2
Label: Verified=0