Update patch set 4
Patch Set 4: Code-Review-1

Since we are having workspace-lock directly in QMakeProjectInfo.updateActiveConfiguration() the deadlock happens more frequently. Right now the lock is acquired on each call - previously it happened only when the project-description was not loaded yet.

This basically means that call to CoreModel.getDefault().getProjectDescriptionManager().getProjectDescription() needs to be called outside of any 'sync' lock anywhere.

Patch-set: 4
Label: Code-Review=-1