Update patch set 3
Patch Set 3:

Thanks for the patch William, it is very good.  I'm finishing off the review but I wanted to mention that I also thought we didn't need IDisassembly3, but then, when reading your comment about allowing others to possibly improve performance, I think it is a nice thing to have.  Let's keep it, even if we don't use it.  As it will be part of DSF itself, it can prove useful to others.

Patch-set: 3