Update patch set 9
Patch Set 9:

In response to the two general comments on Patchset 4, 
please see my answers below:

1- If I launch a CDI session, right-click on the reg view then start a DSF session, the register actions are always disabled in DSF. Vice versa is also true. If I close and re-open the reg view, I can reset the behavior, and re-trigger it.

True, this was a scenario I overlooked, this is no addressed in patchset 9


2- In CDI, the Remove action and Edit action in the reg view, are now grayed-out when selecting a register. They used to not be shown at all. Do you know why that change happened? It is always better not change such behaviours without a good reason.

I agree we should avoid impacts to existing functionality, 
however I have not found a good way to achieve no impacts to CDI, 
I.e. The CDI contributions for Group Edit/Remove were based on an ObjectContribution associated to IPersistableRegisterGroup, 
This interface is applicable to CDI only, We are lacking an interface common to DSF and CDI as the CDI model relies on the Debug model wheras DSF-GDB does not. 
  To make things more interesting this code resides at debug.ui which is a base to DSF-GDB, therefore does not have knowledge of our DSF context structure. 
  So I decided to define them as actions common to both which are present when debug.ui is activated.
With the minor impact of being present and disabled when not selecting a corresponding group.

  Any ideas are very welcome !

Patch-set: 9
Label: Verified=0